Vim 9.0 related fix: `filetype plugin on`
This commit adds these lines to the .vimrc
filetype plugin on
filetype indent on

At the very least, the `plugin` line seems to be necessary to get the
same behavior that I had when using this .vimrc with Vim 8.2.

Notably, it would load `/usr/share/vim/vim82/ftplugin.vim` without said
line in Vim 8.2, but in Vim 9.0 it does not, by default, load
the corresponding `/usr/share/vim/vim90/ftplugin.vim` file.
But now, with the `filetype plugin on` line added, it will load that
file.

This was noticable because hitting Enter (to create a new line) while
inside a comment would no longer automatically insert a comment leader
like it used to. Now it should do that across both versions.